refactor: changed query and dto fields

+++ src/Application/AccommodationService.Application.Models/Rooms/Room.cs
@@ -1,3 +1,3 @@
 namespace AccommodationService.Application.Models.Rooms;
 
-public record Room(long RoomId, long HotelId, long RoomNumber, RoomType RoomType, decimal Price);
+public record Room(long RoomId, long HotelId, int RoomNumber, RoomType RoomType, decimal Price);
