+++ src/Application/AccommodationService.Application/Dtos/RoomDto.cs
@@ -2,4 +2,4 @@ using AccommodationService.Application.Models.Rooms;
 
 namespace AccommodationService.Application.Dtos;
 
-public record RoomDto(long RoomId, long HotelId, long RoomNumber, RoomType RoomType, decimal Price);
+public record RoomDto(long RoomId, long HotelId, int RoomNumber, RoomType RoomType, decimal Price);
